requirements.txt
neurosity~=2.1.1
phue~=1.1
python-dotenv~=1.0.1
pygame~=2.6.0
pyserial~=3.5
numpy~=1.26.4
scikit-learn~=1.5.0
scipy~=1.13.1
mindrove~=3.0.0